buck: Applicable files for includes
Summary:
Diff D85429234 titled `[ELP] Fix incorrect project load on go to def in thrift file` checks when a new file is opened in the LSP client if its canonical path is in the list of applicable files for a given project, and of so does not attempt to load a new project for it.

But for historical reasons the list of applicable files did not include `.hrl` ones.  This diff fixes that, together with the change in `elp.bxl` introduced in D85573080 `[buck/prelude] elp.bxl keep full include file name` which returns the actual applicable files rather than just the directory containing them.

Reviewed By: TheGeorge

Differential Revision: D85573144

fbshipit-source-id: da3eb35f1d80e2107ce4ec9eb9585db710aec7ac