buck: improve target classification for prelude vs third-party
Summary:
This diff operated in the context of the buck project model.
Up to now, the buck erlang prelude files had not been updated to be eqwalizer error-free.

They are now fully eqwalized. Part of the internal usage is to tweak both the common test prelude files and our internal source, and so it is convenient to report eqwalizer diagnostics, via `elp eqwalize-all` for these as well.

The buck project model has historically classified the prelude files as third-party, which maps them to a `TargetType::Dep`, and so eqwalizer does not process them.

This diff changes the logic, so that if any of the `.elp.toml` included targets for a project include a buck prelude target, then it will classify all prelude target files as `TargetType::App` instead.

This has the result of ensuring that they are processed for `elp eqwalize-all` of that project.

Reviewed By: jcpetruzza

Differential Revision: D85761256

fbshipit-source-id: 39a8a56870662d8c148f5a646c7ae63a262693a4