Improve the performance of elp ssr
Summary:
The `elp ssr` feature was initially landed as a quick proof of concept.

In paractice, it runs much slower than it should, because it is delegating to the machinery that runs `elp lint`, which does a lot more work than is needed.

This diff makes `elp ssr` a completely standalone command, which only calls the specific diagnostic checker for an ad hoc SSR diagnostic.

Reviewed By: jcpetruzza

Differential Revision: D85760580

fbshipit-source-id: 5bd38f78b08614be99c4a7d4cc69ebed8832fc70