use jar version for unoptimised builds
Summary:
This changes local debug builds (linux only) to use a `jar` rather than building the native image for eqwalizer. This speeds up iteration significantly.

We had this before, but struggled with java versions. In this iteration we just use tha java from graalvm that we would use to build the native image - that way we have a consistent java version that we fully control and don't depend on the OS.

Reviewed By: TD5

Differential Revision: D85784216

fbshipit-source-id: 188dd7207865fda770bf2e154e18049030acdc29