BE: tree print with ranges
Summary:
When developing a feature in ELP it is often useful to tree print a body, to see clearly what the HIR representation is in it.

This is disconnected from the original source, however, as we have no indication of precisely where each part originates.

This diff updates tree print to optionally take a `BodySourceMap`, and if present it prints the range of the item when introducing the constructor.

It is use to add a test showing that the lowering of Expr::Paren is incorrect.

Reviewed By: TD5

Differential Revision: D85866631

fbshipit-source-id: 866eeba845f396ba51fcfd01e92d9c264e450534