CLI ssr: allow specifying macro and parens strategy
Summary:
Structural Search (SSR) matches the syntax tree used internally in ELP, called HIR. This is conceptually similar to Erlang Abstract Forms.

When we lower the surface Erlang language into this, we expand macros and strip of parentheses, as the precedence of items is captured in the actual HIR representation.

But sometimes when we are searching over Erlang code we want to explicitly match parentheses, e.g. if we want to find all instances of redundant duplicate ones.

This diff allows you to do this by issue the command

```
elp ssr --parens "((_A))"
```

Likewise, sometimes you are interested in occurrences of macros. This can now be provided using a match like

```
elp ssr --macros no-expand "?BAR(_X)"
```

Reviewed By: TD5

Differential Revision: D85876784

fbshipit-source-id: ba62d83ed8ea9c7ac30cceda2cac446ecdeefb06