Properly process changes to open configuration files
Summary:
ELP as an LSP server operates in a dyamic environment, by design. The key thing ts must deal with is changes to files.  A given project loaded into ELP has to process file-changes from two primarily different sources.

The first is open files in the LSP client (VS Code), and these are sent via LSP protocol as they happen, as well as generating a `DidSaveTextDocument` notification when saved.

The other is changes that happen on the filesystem directly, from any source, such as an external editor, version countrol operation, build, or the like. These are monitored via a file watcher, and the watching is also delegated to the LSP client.

There are some wrinkles to this.  We send a registration message to the client to watch all files of interest in a project, both source and configuration.  So of necessity this will overlap with any files that are also open in the IDE.  To avoid double processing these, we only process open files in the `DidSaveTextDocument` notification, and do not process open files in the `DidChangeWatchedFiles` notification.

This is all very well, but at some point there was a rationalisation of change processing, with the realisation that we only care about `DidSaveTextDocument` notifications for triggering diagnostic requests for open files for diagnostic sources not providing on-change diagnostics.  Which is now most of them.

So we modified the registered scope for `DidSaveTextDocument` to only be `.erl`/`.hrl` files.

This meant that we no longer processed changes to configuration files that are open in the IDE, which is the main way they pick up changes.

## This diff

We restore the desired immediate reaction to changed files by never doing the check in the `DidSaveTextDocument` notification processing, since it never applied anyway, and so always doing it in `DidChangeWatchedFiles`.

We also tidy up the registering of watched files by first unregistering, otherwise the client accumulates multiple watches for the same files, and sends a change notification for each registration it has.  Even with this removal, we still get 2 notifications per file change, so add logic to de-dupe them while processing, as updating config can be an expensive operation.

Reviewed By: jcpetruzza

Differential Revision: D85956439

fbshipit-source-id: 961618711347d173596f69f5fd94a435b3fd45b1