BE: include diagnostic code in test fixtures
Summary:
Our declarative test fixtures allow us to annotate the fixture with the epxected diagnostics to be produced by the test.

Up to now, this has included the severity and text only.

This (massive but regular) diff updates this to include the diagnostic code too.

This makes it clearer that the code is correct, and makes it easier to search for examples of its usage in the test code.

Reviewed By: robertoaloi, TD5

Differential Revision: D86103082

fbshipit-source-id: 6e4ede050350cc90aa437f68da0900a66b7d0625