buck2: Improve pop-up when the buck project is malformed
Summary:
When ELP is loading a buck project it does queries using buck.
Sometimes these fail, because the current state of the buck configuration files is invalid.
In that case, we show a pop-up suggesting that the user examine the buck dashboard, and provide a URL for this.

It is not always clear to the user that this is related to a malformed buck project, rather than a bug in ELP.

This diff tries to make it clearer.

Note that we unfortunately do not have the ability to use any kind of formatting in the displayed message.

Reviewed By: TD5

Differential Revision: D86510447

fbshipit-source-id: 57a93f829d37a8a4d7c8be88d6cf68adcf41c859