1/n: includes: add not-yet-failing test case for nested include buck targets
Summary:
This diff sets up a test case to demonstrate a problem which will be fixed by this stack, as it is currently silently ignored.

As such, there is no failure reported in the test. That will change later in the stack.

It is possible to configure a buck2 project with nested include files so that it compiles with buck2, and also with our erlang_service, but that we are unable to properly lower it into HIR and so do not resolve macros in ELP.

This occurs when we have three buck2 targets, A,B,C each in a separate directory.
- A has both B and C as dependencies, but
- B does not have C as one.

Then
- target A contains a .erl file that includes a .hrl from target B
- the .hrl in target B includes a .hrl from target C

So from A we can resolve the entire chain of includes via dependencies, but from B we cannot.

And ELP lowers the file into HIR for B in the context of B, so does not resolve the include at C, and silently fails to do so, entering a `Expr::Missing` value instead.

Reviewed By: TD5

Differential Revision: D86963334

fbshipit-source-id: 57b51714513fc78cd21083adea6abd12a9350571