Convert undocumented_module linter to use trait
Summary:
Even if the `undocumented_module` is supposed to be disabled by default, it still runs. This is due to a flawed logic in `diagnostics.rs`, originally introduced in D59272525.

This is problematic, since open-source projects such as OTP see a huge number of errors when running the `elp lint` command without specifying a custom configuration.

By converting the linter to use the new trait, we incidentally fix the bug.

One more linters uses a `default_disabled: true` property in its descriptor: `undocumented_function`. It will be converted next, and the logic to handle the `default_disabled` will eventually be removed.

Reviewed By: alanz

Differential Revision: D87059857

fbshipit-source-id: 7fb6eff8773c786c6894548c914b8e3c78c38745