Convert the undocumented_function linter to use trait
Summary:
Even if the `undocumented_function` is supposed to be disabled by default, it still runs. This is due to a flawed logic in `diagnostics.rs`, originally introduced in D59272525.

This is problematic, since open-source projects such as OTP see a huge number of errors when running the `elp lint` command without specifying a custom configuration.

By converting the linter to use the new trait, we incidentally fix the bug.

This was the last linter using the `default_disabled: true` logic. At this point the enabled/disabled logic can be simplified.

Reviewed By: alanz

Differential Revision: D87060203

fbshipit-source-id: 1121d286c766fb461420e3268aca817169020d7d