Convert missing_compile_warn_missing_spec linter to use trait
Summary:
Even if the `missing_compile_warn_missing_spec` is supposed to be disabled by default, it still runs. This is due to a flawed logic in `diagnostics.rs`, originally introduced in D59272525.

This is problematic, since open-source projects such as OTP see a huge number of errors when running the `elp lint` command without specifying a custom configuration.

By converting the linter to use the new trait, we incidentally fix the bug.

Two more linters use a `default_disabled: true` property in their descriptors: `undocumented_module` and `undocumented_function`. They will be converted next, and the logic to handle the `default_disabled` will eventually be removed.

Reviewed By: alanz

Differential Revision: D86978483

fbshipit-source-id: 5a89ff0fd1297a1e75db9d4b60d75b70f98d1e0c