BE: track related info in declarative tests
Summary:
Our declarative tests allow us to add annotations for expected diagnostics, so we can clearly see and check what is generated.

But at present there is no way to check any related information for a diagnostic.

This diff adds that capability, showing related information together with the FileId and range each refers to.

Reviewed By: robertoaloi

Differential Revision: D87316965

fbshipit-source-id: ea7db2980b11950c3552befec7b666665ceef84b