add file_id to diagnostic RelatedInformation
Summary:
A given diagnostic can have related information in it, which carries relevant context to the provided diagnostic.
For example, for a head mismatch diagnostic, it reports the specific head location(s) that do not match.

The LSP spec uses a `Location` for these, which includes a file URL, meaning the related info can be in a different file from where the diagnostic is reported.

In ELP, we currently do not provide a way of populating this information, so can only refer to related information within the same file.  Thic can result in erroneous outcomes.

## This diff
We include a `FileId` in our internal `RelatedInformation` structure, and map it to a URL when serializing for transmission to the LSP client

Reviewed By: robertoaloi

Differential Revision: D87316542

fbshipit-source-id: 5f94eb4fe5bb73e7bf1fff7cad458231c51bd6f0