BE: fix off-by-one in `elp lint` diagnostic output
Summary:
When running any elp CLI command that produces diagnostics, the locations are incorrect, as they use the internal coordinate frame rather than the usual one for IDEs and other tools.

This means each line and col location is one less than it should be.

This diff fixes that, so the locations make sense to users.

Note: it does not affect output using `--format json`, which already does this.

Reviewed By: TD5

Differential Revision: D87539540

fbshipit-source-id: 27abd4ac5fb9e5056fc9e3397257d9e5cb9e678d