Allow severity method to receive &Semantic and FileId as arguments
Summary:
# Context

We want to convert the `edoc` linter to use a trait.

# Problem

The `edoc` linter has a different `Severity` for source code and test files.

# Solution

We can extend the `severity` and `cli_severity` methods to accept `sema` and `file_id`. This way, we can customize the severity based on the file. The new arguments will be used by the edoc linter in a subsequent diff.

Reviewed By: alanz

Differential Revision: D87545262

fbshipit-source-id: d7546d7d77ec019c2a2edc846e69134e93a6fe90