Move tree-diff functionality from syntax to ide_db crate
Summary:
# Context

The `text_edit` crate contains a single file, `lib.rs`. We can remove the entire crate by moving the functionality into a `text_edit.rs` in the `ide_db` crate. This is in line with Rust Analyzer.

# Problem

The move would eventually create a circular dependency between `syntax` and `ide_db`.

# Solution

To untangle the circular dependency, we can move the `tree-diff` functionality to the `ide_db` crate. Again, this is in line with Rust Analyzer.

Reviewed By: alanz

Differential Revision: D87524148

fbshipit-source-id: 90b1df3d4191a63c0d74c49af9fafec6578f530c