ssr: Improve rendering of result
Summary:
Inside ELP an SSR match is exposed as a diagnostic.  So the way it is output to the command line is as a diagnostic, giving location and the pattern that matched.

This is not very useful when exploring a code base from the command line.

This diff brings the ability to show the source for a matching diagnostic.  The command line interface is modeled on the one for grep, so it should be familiar to users.

It will automatically use colour if in a TTY context, but this can be controlled via CLI or environment variable.

Reviewed By: TD5

Differential Revision: D87634862

fbshipit-source-id: 52604d21100274b2abba9e364acbb972914aa4ac