ssr: stream command line match results
Summary:
It is annoying to wait for the entire body of code to be processed before seing results. Since the ssr search is based purely on the ast of a module, the search is embarrasingle trivial.

So report the results as they are found.

This does mean the summary statistic has to come at the end.

Reviewed By: TD5

Differential Revision: D87651432

fbshipit-source-id: b51b7bd94f8069113154adac9b1bf68f9b7351b3