Remove unclear bits of W0050 docs
Summary:
# Context
W0050 asks users to prefer the more efficient `tuple_size/1` and `byte_size/1`, instead of `size/1`.

# Problem
The explanation is muddled by some advice to always guard the usages of `byte_size/1` with `is_binary/1`. This doesn't make intuitive sense and there's nothing in the docs of `byte_size/1` that suggests this is advisable in any way.

# This diff
Remove this part. I'm also removing the part that rewrites `size/1` in terms of `byte_size/1` and `tuple_size/1` as it doesn't seem to be adding anything.

Reviewed By: TD5

Differential Revision: D87986348

fbshipit-source-id: dbd6aeec71989bc367f3d22b58a71d893b7fff76