Do not check for eqwalizer disappearing if we did not unpack it
Summary:
The normal case for the ELP executable is to have a copy of eqwalizer as a Graalvm exe embedded inside ELP. At startup, this is unpacked to a temporary directory, so we can run it.

In the past we had issues where external processes would clean up the temporary directory, and so it would disappear. To mitigate this, we check if the file still exists before running it, and recreate it if not.

BUT, if we are running with ELP_EQWALIZER_PATH pointing to the eqwalizer jar file, as per the OSS local build instructions, then we do not unpack it to a temporary directory, *and* the command name is set to simply "java", which is not a file, as it should be resolved from the path. As a result the current existence check fails, and we continuously report that eqwalizer has disappeared.

This diff changes the check to only be made if we have in fact unpacked eqwalizer to a temporary directory.

Reviewed By: TD5

Differential Revision: D88071028

fbshipit-source-id: 21e7c4897dd6a8f6ae6fe137c413b2fab094d936