Correctly report diagnostics from header file
Summary:
It can happen that an included file has errors in it.

ELP is supposed to notice this, and report it with an L0000 (issue_in_included_file) diagnostic.

Unfortunately the logic for reporting this in the erlang service did not properly account for the exact location information reported in the `file` attribute, so this always fell back to a location of 0,0 in the including file. As a result, the diagnostic did not show up properly.

## This diff

We fix it by reporting the location according to the representation  used in the erlang service, and making sure we choose the correct include location.

We also add the diagnostic found in the included file as related information, so it is available in the original file

Reviewed By: robertoaloi

Differential Revision: D88010399

fbshipit-source-id: 4a1d2d76a809c104db059e9433200e1bfcf2ac59