Include related information in `elp lint` non-json output
Summary:
An ELP Diagnostic can have related information in it, which provide additional context to the reported Diagnostic.

Although these are shown in the IDE when running as a server, they are not currently shown from the  command line.

This diff updates the `elp lint` command to display them, when not emitting JSON.

There are 3 variations, depending on where the related information is
-  same file: just show the range
- some erlang module: show the module name and range
- an erlang include file: show the full path and range

Reviewed By: robertoaloi

Differential Revision: D88010319

fbshipit-source-id: 26748ec7183a139a57ae97b75a9a61f62da27000