Make it possible to run buck tests in isolation
Summary:
The buck2 tests seem to rely on a global state, which is only set for some of the test cases.

As a consequence, running all tests together works, but running some of them in isolation fails.

Consistently set the global state in all tests. As a follow-up, we could refactor this in a helper function.

Reviewed By: alanz

Differential Revision: D88274947

fbshipit-source-id: 5f22894e74bd41458139a93cd063647d95b86296