Fix false positive for undefined_function linter
Summary:
It can happen that, when looking for undefined functions, we could not resolve the module name at compile-time.

With the old logic, we would incorrectly flag the case as an "invalid function".

With the improved version, we do not flag the case if we are not sure, avoiding a number of false positives.
A test case highlighting the issue is added as a unit test.

Reviewed By: TD5

Differential Revision: D88474342

fbshipit-source-id: f0f5ae1272f8264c5e11c6e8a7855e75a2832841