streaming results for elp lint
Summary:
At the moment the `elp lint` command processes all the files, and then renders the diagnostics at the end.  This can be annoying if you are wanting to just check for occurrences of a diagnostic.

This diff changes the behaviour to stream the results instead as they arrive.
This behaviour is turned off if
- the new `--no-stream` argument is given
- `elp lint` emits json formatted diagnostics
- `elp lint` applies fixes

Reviewed By: TheGeorge

Differential Revision: D88010824

fbshipit-source-id: dc725c1d503f56da6bb1c392cd88f635832320dc