BE: report malformed test fixture
Summary:
A declarative test fixture can have multiple files in it, in which case each has a metadata line giving the file name. But we have not been checking if it is omitted from the first file, leading to confusing behaviour when writing tests.

This diff updates test fixture parsing to panic in that case, with an appropriate message.

Reviewed By: TD5

Differential Revision: D88947477

fbshipit-source-id: 7ef7151fafb44d087b1da25c55454f20899ca67f