Allow project discovery for OTP
Some clients, such as emacs eglot, do their own project discovery when
a new file is opened, even from the target of a go to definition.

They then launch a fresh ELP server if they determine a different
project root.

In the ELP server setup we preconfigure the OTP project root as known,
but with a project manifest of None. This means go to def into an OTP
file ends up with an uninitialized project in the new server.

This diff makes it so that if the project root is known, but has no
project manifest, we do discovery on that root.