BE: for a fold, track the macro definition for ctx.in_macro
Summary:
When folding with visible macros, we keep track of a macro stack, so we can decide how to process diagnostics within the body of a macro expansion.

To date we have simply stored the `HirIdx` of the macro call, so it can be retrieved and examined by anything that needs it.

But the thing we need when processing a macro is in fact what definition was expanded.  This diff adds it to the information stored for easy access.

It will be used in a subsequent diff.

Reviewed By: TD5

Differential Revision: D89269840

fbshipit-source-id: 7ed6212b658de91cdd3e20701edb51161966e8fa