7/n: rename module: check module name in a call
Summary:
When identifying a call from a reference, we start from the ast token for the module name, and ascend the ast until we find a call.

But up to now we have not checked that the module we started from is the module part of the call MFA.

This diff adds that check.

Reviewed By: robertoaloi

Differential Revision: D88934411

fbshipit-source-id: 81a57520ec834210fcea8fd6920653cc45f9890c