5/n: rename module: move original file
Summary:
The final step in renaming a module.  Move the old file, and apply the needed edits to it.

This diff also simplifies the mechanics of the rename, to be more general and in line with other renames, by iterating over usages of the module, rather than ad-hoc checks for things that may be needed.

It is still ad-hoc, in the sense that it only looks for call-like things at the moment (fully qualified calls and types). Other usages will come in later diffs

Reviewed By: robertoaloi

Differential Revision: D87861050

fbshipit-source-id: d393638b310a257f51151fc3caf1b1445c856700