8/n: rename module: well known function arguments
Summary:
Certain functions are known to take a module name as an argument.  We already have a list of those that correspond to referencing a function as an MFA, in `to_def.rs`.  We extend this with some other examples, and check for renaming them too as part of the module rename.

The list is currently incommplete, and can be filled out based on usage. A future diff may look at ways of generating it from `-spec` information.

Reviewed By: robertoaloi

Differential Revision: D88943301

fbshipit-source-id: 05c67a108b1dc5b64f4353872427361fd039171f