Enable undefined_function linter in CI (prod only)
Summary:
The linter is enabled by default in ELP, and warnings appear in the IDE.
But the same linter is not explicitly added to the CI configuration, so warnings are not reported on diffs or in Contlint.

This diff enables the linter, aligning the IDE and the CI experience. It also makes it possible to track violations via ContLint.

We exclude known false positives via config.

Ideally, we would like to report undefined functions in all files, but there are too many false positives in test files to do so.
This is often due to mocked modules and test suite cleverness. We can revisit this decision in the future. See T249044930 for details.

We disable the linter for tests both in the default value (for the future) and the config (so we don't need to wait for a ELP release to land this change).

Reviewed By: alanz

Differential Revision: D89393641

fbshipit-source-id: bfb9ac760ba79566301087033af95bb805fdeaa0