Do not run the old_edoc_syntax linter against test files by default
Summary:
Documentation is generally not produced for test files, so do not run the `old_edoc_syntax` linter for tests. If not, edoc references in tests should be converted to plain comments, not to `-moduledoc` or `-doc` attributes.

This behaviour can always be overwritten by setting the `severity` or the `include_tests` properties of the linter via config.

Reviewed By: alanz

Differential Revision: D89374552

fbshipit-source-id: 8a4299fcbd16a82f62efbc14f47fef46cd08fbc0