1/n: bound var in match: detect problem
Summary:
In non-test Erlang code, the following is almost always intended as assignment of a value to the LHS variable

```
Var = expr()
```

It is actually a match, and may fail if `Var` is already bound, and the new value differs from the old.

This diff introduces a diagnostic to warn of this case.

Reviewed By: robertoaloi

Differential Revision: D89374198

fbshipit-source-id: 19dcee50ad6e71d0a606a0eb6cf548b7194d61e6