Only enable Buck2 project model tests if feature is enabled
Summary:
We already have a feature to understand if Buck2 is enabled or not. Let's use that one instead of a constant.

This change has no effect anyway since the only two use cases are currently marked as "ignored".

Also remove the extra check inside the helper (we already do it in both tests).

Reviewed By: alanz

Differential Revision: D89549782

fbshipit-source-id: 82e5bbf00c463dca3f4937d2d2a12d16aebc3c19